:wrench: Make more indentation rules based on line of first sibling (#15)
This is a continuation of some previous work to make indentation rules
more contextual. Indentation styles tend to differ depending on
whether the first child of the parent is on the same line as the
parent or on its own line.

This commit extends this logic to tuple expressions, vector and matrix
expressions, type parameter lists, and curly brace expressions.

This commit also corrects the behavior of assignment expressions.
Those were incorrectly aligning to the identifier being assigned to
when the julia-ts-align-assignment-expressions-to-first-sibling
variable was set to t.

Lastly, this commit generalizes the matcher function used to implement
these features and expands its documentation.