:wrench: Simplify indentation rules
Instead of using a function that receives and argument to verify if a
sibling is on the same line or not of its parent, it is better to
provide two different functions. Notice that we do not need to change
the type of the verification with respect to some runtime value. Hence,
we can create two different functions.